Separate test.js module
